Switch from Rust to Python
I initially chose Rust because I believed in "maximum performance." But,
it wasn't easy. Futhark doesn't generate a nice wrapper library for Rust
like it does for Python, so I had to use bindgen. Tricky, but at least
I learned something new. Handling video, though, was a big pain. FFmpeg
and GStreamer seemed like the only options. But, the number of FFmpeg
crates worried me. It seemed like there were a lot of experiments, but
no canonical crate. The crates also seemed pretty low-level, and my eyes
glazed over at the (understandable) complexity of FFmpeg's API. I then
looked at GStreamer, but the amount of code required to setup GObject
(just from the rgb2gray example) turned me off.

I'm sure this could be done with Rust, but I want to focus on the actual
seam carving, and not the intricacies of handling video. PyAV should
make that much simpler. Python is also nice because Futhark can create
Python libraries, as mentioned earlier. Finally, performance shouldn't
be too much of an issue, since the Python code should just be shuttling
data back and forth between FFmpeg and the GPU. If there is a cost, I
hope it's not more than a constant factor slowdown. Anyway, this is not
a serious system, so I'm happy as long as it runs reasonably quickly.