Use 2D array for energy calculation
It's about 3-4x faster, probably because we don't have to perform a
division and modulo for each pixel. I'm not sure why I switched to a 1D
array when my initial experiments used 2D arrays. Futhark shouldn't have
trouble optimizing since we're not expecting to operate on tiny frames.
Maybe because it was easier for FFI? But there's a 2D array type. It's
not as easy as NumPy, but it's not harder than wrapping 1D arrays.