Add index map and sum seam energy functions
This version of sum_seam_energy has similar performance to Athas's
version while supporting array sizes that are not powers of two. So, I
think it should be pretty good. Not sure if there's a better way to
write index_map. Perhaps some crazy shift-and-compare method can bring
it down to two comparisons per pixel?