Basic video carving functionality
The program can now carve a video by processing each frame separately.
However, there are some problems:
- Performance: It feels very slow. That's expected, but the question is
whether it's due to the implementation, algorithm, or hardware. The
latter two I can't fix, but if it's the implementation, then more work
needs to be done.
- Crashes: Carving too many pixels (i.e. more than 4) on a 720p video
leads to a crash. My computer freezes and the fan kicks up. When
everything is responsive again, the program is stuck at a frame with the
CPU still going at 100%. I'm not sure where it stops or why it gets
stuck. Maybe the GPU is overloaded or overheats and dies? I'm not sure
why the Python program doesn't crash then.