Randomly select minimum seam
In 0253df7, I said that "<= vs < shouldn't make a difference in
[min_seam_index's] reduction". This turned out to be true, but there was
a bigger problem: 0 is not actually a neutral element. If there's a
frame where all seams have the same energy (e.g. a solid background), 0
will always be picked. This means that all seams will bunch up on the
left, and the following frames will also have all their seams on the
left due to temporal coherence.

So, if there is more than one seam with the same energy, we have to
randomly pick one of them. This means that finding the min seam index is
now done on the CPU. I don't know what the performance impact is, but I
think it should be negligible. The Numpy code is microseconds.
Transferring over a row of seam energies instead of a scalar is
technically more data, but a few thousand values vs one value is nowhere
near the bandwidth. The synchronization cost likely dominates.