Switch from progress to tqdm
I said in 97e1ef8 that other progress bar libraries looked complicated
compared to progress. tqdm was one of these "other libraries". But I
only thought it was complicated because it had extensive docs. At a
glance, it looked like overengineering. But, the library is just
actively maintained and widely used. tqdm just works and doesn't need
fixes like the atexit hook, so I think it's a better choice.