Switch from PyOpenCL to futhark-ffi
Our code is a prime example of launching lots of small kernels. This is
bad in general, but PyOpenCL just adds more overhead on top of that.
Simply switching to futhark-ffi is 2.15x faster. Fetching the index map
and seam energy at the same time instead of trying to "hide latency" is
1.13x faster on top of that. I'm not exactly sure why that is. Maybe
there's hidden synchronization or the GIL is kicking in or something. To
successfully pipeline the transfer, you'd probably need to go to the
OpenCL level and do things manually.