Use 64-bit sizes
This is required as of Futhark 0.18.1. The benchmarks aren't very
rigorous, but there is a slowdown: (We use the average time on ten runs
of the 4000x4000 test case. Values are new time/old time.)

  CPU:
    saliency:       0.977
    horz coherence: 0.938
    index_map:      1.284
  GPU:
    saliency:       1.368
    horz coherence: 1.030
    index_map:      1.136

The faster times on CPU are probably due to a fluke. Overall, I'm not
sure why the results are what they are. If branch divergence is a major
factor, the random data probably affects the results a lot.
I'm guessing that performance could be improved by avoiding indexing,
but I'm not sure how to perform these sorts of computations without
using explicit indexing. Maybe if there was a "windowed map" primitive.