Use i16 indexes when possible
It turns out that the slowdown from 64-bit sizes is worse than I
thought. Just switching to i16 when possible gives a decent performance
boost by reducing memory usage--both on the GPU itself and in transfers
between host and device.
The only annoying part is that `iota` and `replicate` only return i64.
So, I needed to copy the functions from the standard library and then
use size coercions. This should be faster than materializing the array
and then using `map i16.i64`.